Separate HomePage class from main
